+++ src/main/java/pfs/lms/enquiry/appraisal/projectlocation/SubLocationDetailResource.java
@@ -22,13 +22,13 @@ public class SubLocationDetailResource {
     private String district;
     private String region;
     private String nearestVillage;
-    private Integer nearestVillageDistance;
+    private Double nearestVillageDistance;
     private String nearestRailwayStation;
-    private Integer nearestRailwayStationDistance;
+    private Double nearestRailwayStationDistance;
     private String nearestAirport;
-    private Integer nearestAirportDistance;
+    private Double nearestAirportDistance;
     private String nearestSeaport;
-    private Integer nearestSeaportDistance;
+    private Double nearestSeaportDistance;
     private String nearestFunctionalAirport;
-    private Integer nearestFunctionalAirportDistance;
+    private Double nearestFunctionalAirportDistance;
 }
